custom exceptions moved to api module
