+++ api/src/main/java/com/tinqinacademy/hotel/api/base/OperationProcessor.java
@@ -1,6 +1,6 @@
 package com.tinqinacademy.hotel.api.base;
 
-import com.tinqinacademy.hotel.api.error.Errors;
+import com.tinqinacademy.hotel.api.exceptions.Errors;
 import io.vavr.control.Either;
 
 public interface OperationProcessor<I extends  OperationInput, O extends OperationOutput > {
