+++ api/src/main/java/com/tinqinacademy/hotel/api/exceptions/Errors.java
@@ -1,4 +1,4 @@
-package com.tinqinacademy.hotel.api.error;
+package com.tinqinacademy.hotel.api.exceptions;
 
 import lombok.*;
 import org.springframework.http.HttpStatus;
