+++ api/src/main/java/com/tinqinacademy/hotel/api/exceptions/custom/CustomException.java
@@ -1,4 +1,4 @@
-package com.tinqinacademy.hotel.core.exceptions.custom;
+package com.tinqinacademy.hotel.api.exceptions.custom;
 
 import org.springframework.http.HttpStatus;
