+++ core/src/main/java/com/tinqinacademy/hotel/core/exceptions/ExceptionService.java
@@ -1,6 +1,6 @@
 package com.tinqinacademy.hotel.core.exceptions;
 
-import com.tinqinacademy.hotel.api.error.Errors;
+import com.tinqinacademy.hotel.api.exceptions.Errors;
 
 public interface ExceptionService {
     Errors handle(Throwable throwable);
