+++ core/src/main/java/com/tinqinacademy/hotel/core/operations/hotel/BookRoomOperationProcessor.java
@@ -1,13 +1,13 @@
 package com.tinqinacademy.hotel.core.operations.hotel;
 
-import com.tinqinacademy.hotel.api.error.Errors;
+import com.tinqinacademy.hotel.api.exceptions.Errors;
 import com.tinqinacademy.hotel.api.operations.hotel.bookroom.BookRoomInput;
 import com.tinqinacademy.hotel.api.operations.hotel.bookroom.BookRoomOperation;
 import com.tinqinacademy.hotel.api.operations.hotel.bookroom.BookRoomOutput;
 import com.tinqinacademy.hotel.core.exceptions.ExceptionService;
-import com.tinqinacademy.hotel.core.exceptions.custom.BookedRoomException;
-import com.tinqinacademy.hotel.core.exceptions.custom.BookingDatesException;
-import com.tinqinacademy.hotel.core.exceptions.custom.NotFoundException;
+import com.tinqinacademy.hotel.api.exceptions.custom.BookedRoomException;
+import com.tinqinacademy.hotel.api.exceptions.custom.BookingDatesException;
+import com.tinqinacademy.hotel.api.exceptions.custom.NotFoundException;
 import com.tinqinacademy.hotel.core.operations.BaseOperationProcessor;
 import com.tinqinacademy.hotel.core.utils.LoggingUtils;
 import com.tinqinacademy.hotel.persistence.model.entity.Booking;
